+++ test/unit/api/c/capi_grammar_black.cpp
@@ -36,6 +36,7 @@ class TestCApiBlackGrammar : public ::testing::Test
   void TearDown() override
   {
     cvc5_delete(d_solver);
+    cvc5_term_manager_release(d_tm);
     cvc5_term_manager_delete(d_tm);
   }