+++ test/unit/api/c/capi_symbol_manager_black.cpp
@@ -34,6 +34,7 @@ class TestCApiBlackSymbolManager : public ::testing::Test
   {
     cvc5_symbol_manager_delete(d_sm);
     cvc5_delete(d_solver);
+    cvc5_term_manager_release(d_tm);
     cvc5_term_manager_delete(d_tm);
   }
 
@@ -53,6 +54,7 @@ class TestCApiBlackSymbolManager : public ::testing::Test
     Cvc5Command cmd = cvc5_parser_next_command(parser, &error_msg);
     ASSERT_NE(cmd, nullptr);
     (void)cvc5_cmd_invoke(cmd, d_solver, d_sm);
+    cvc5_parser_release(parser);
     cvc5_parser_delete(parser);
   }